+++ resources/views/procurement/asset-purchase-request/admin/_index_filtered.blade.php
@@ -7,7 +7,9 @@
             <th class="erp-th text-center">Title </th>
             <th class="erp-th text-center">Remarks </th>
             <th class="erp-th text-center">Status </th>
-            <th class="text-end erp-th">Action</th>
+            @if(hasPermission('manage-asset-product-purchase-request'))
+                <th class="text-end erp-th">Action</th>
+            @endif
         </tr>
         </thead>
         <tbody class="erp-tbody">
@@ -30,21 +32,22 @@
 
                         <h4 class="text-center d-table-title {{strtolower($pr::REQUEST_STATUSES[$pr->request_status])}}-status">{{ $pr::REQUEST_STATUSES[$pr->request_status] }}</h4>
                     </td>
-
-                    <td class="text-end erp-tbody-td">
-                        @if($pr->request_status == $pr::REQUEST_STATUS_NEW)
-                            <div class="erp-action-t">
-                                <div class="dropdown dropdown-action">
-                                    <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
-                                    <div class="dropdown-menu dropdown-menu-right">
-                                        <a class="dropdown-item" href="javascript:void(0);" onclick="getRequestDetails({{$pr->id}})"><i class="fa-solid fa-circle-info m-r-5"></i> View Details</a>
-                                        {{-- <a class="dropdown-item" href="{{ route('procurement.admin.asset-purchase-request.edit',$pr->id) }}"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a> --}}
-                                        <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('procurement.admin.asset-purchase-request.delete',$pr->id) }}', 'reloadAjaxGetData') "><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
+                    @if(hasPermission('manage-asset-product-purchase-request'))
+                        <td class="text-end erp-tbody-td">
+                            @if($pr->request_status == $pr::REQUEST_STATUS_NEW)
+                                <div class="erp-action-t">
+                                    <div class="dropdown dropdown-action">
+                                        <a href="#" class="action-icon dropdown-toggle" data-bs-toggle="dropdown" aria-expanded="false"><i class="material-icons">more_vert</i></a>
+                                        <div class="dropdown-menu dropdown-menu-right">
+                                            <a class="dropdown-item" href="javascript:void(0);" onclick="getRequestDetails({{$pr->id}})"><i class="fa-solid fa-circle-info m-r-5"></i> View Details</a>
+                                            {{-- <a class="dropdown-item" href="{{ route('procurement.admin.asset-purchase-request.edit',$pr->id) }}"><i class="fa-solid fa-pencil m-r-5"></i> Edit</a> --}}
+                                            <a class="dropdown-item" href="javascript:void(0)" onclick="deleteAjax('{{ route('procurement.admin.asset-purchase-request.delete',$pr->id) }}', 'reloadAjaxGetData') "><i class="fa-regular fa-trash-can m-r-5"></i> Delete</a>
+                                        </div>
                                     </div>
                                 </div>
-                            </div>
-                        @endif
-                    </td>
+                            @endif
+                        </td>
+                    @endif
                 </tr>
             @endforeach
         </tbody>
